created pages folder inside src folder to put all the pages inside the pages folder
