+++ Frontend/src/Components/Navbar/MyNav.jsx
@@ -18,7 +18,7 @@ function MyNav() {
           className="mb-3"
         >
           <Container fluid>
-            <Navbar.Brand href="#">Reciepe</Navbar.Brand>
+            <Navbar.Brand href="#">Recipe</Navbar.Brand>
             <Navbar.Toggle aria-controls={`offcanvasNavbar-expand-${expand}`} />
             <Navbar.Offcanvas
               id={`offcanvasNavbar-expand-${expand}`}
@@ -27,7 +27,7 @@ function MyNav() {
             >
               <Offcanvas.Header closeButton>
                 <Offcanvas.Title id={`offcanvasNavbarLabel-expand-${expand}`}>
-                  Reciepe
+                  Recipe
                 </Offcanvas.Title>
               </Offcanvas.Header>
               <Offcanvas.Body>
