+++ Frontend/src/pages/contactUs/constyle.css
@@ -20,9 +20,12 @@
     
     display: flex;
     flex-direction: column;
-    width: 300px;
+    width: 400px;
     margin: 0 auto;
-    padding: 20px;
+    padding: 40px;
+    
+    
+    border: 5px solid rgba(217, 205, 30, 0.299) ;
     border-radius: 8px;
     box-shadow: 0px 0px 15px 0px rgba(217, 205, 30, 0.299);
     font-family: 'Poppins', sans-serif;
@@ -68,7 +71,7 @@
 }
 
 .contact-form input[type="submit"] {
-    background-color: #606c38;
+    background-color: #ffca3a;
     color: white;
     padding: 10px 20px;
     border: none;
